backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomous-literature-survey"
version = "1.0.0"
description = "Production-grade Agentic RAG for automated academic literature review"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "python-multipart==0.0.20",
    "websockets==14.1",
    # Database
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    # Auth
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "authlib==1.4.0",
    "httpx==0.28.1",
    # AI/ML
    "langgraph==0.2.60",
    "langchain==0.3.13",
    "langchain-openai==0.2.14",
    "langchain-anthropic==0.3.1",
    "langchain-community==0.3.13",
    "openai==1.58.1",
    "tiktoken==0.8.0",
    # Vector store
    "pinecone-client==5.0.1",
    # Cache
    "redis[hiredis]==5.2.1",
    # Validation & settings
    "pydantic==2.10.4",
    "pydantic-settings==2.7.1",
    # Utilities
    "structlog==24.4.0",
    "tenacity==9.0.0",
    "python-dotenv==1.0.1",
    "aiohttp==3.11.11",
    "rapidfuzz==3.11.0",
    "numpy==1.26.4",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "pytest-cov==6.0.0",
    "httpx==0.28.1",
    "ruff==0.8.6",
    "mypy==1.14.1",
    "factory-boy==3.3.1",
    "aiosqlite==0.20.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
strict = false
